Change mp-stacks-metabox to mp-stacks-size metabox
Changed the name of the file for the Brick Size Settings  metabox from being just mp-stacks-meta to mp-stacks-size - also changed the name of related filters used in that file. Additionally, I changed The Size Metabox to use new mp_brick_metabox filter hook. This makes it only load when it needs to.