Text Resize System
This system works to resize text by checking if any p tags are wider than their parent and resizing them if so. This allows people to use huge text sizes for desktop without ever creating overlap on mobile devices. It resizes the text to JUSt fit inside the outer size limit of the screen and/or stack.