Show "Loading Brick..." until fully loaded
Because we do some re-arranging of the Brick editor upon load, to make it look clean, we will hide all page contents and show a "loading brick editor..." message until it is layed out properly.

Also, the message which lets the user know which brick they are editing below the Brick Title on the Brick Editor page now shows on both Add New and Edit Brick pages.