Removed duplicate margin setting
The code was setting the same margin whether it was centered or leftright, but separated by an if stament. This removes the if statement and applies it regardless of the brick alignment.

Also, this keeps the margins if if the brick is less than 600px wide. We were turning them off if they were leftright aligned and less than 600. While this did tighten up content types on mobile, it also threw off any custom margins - like phones that touch the bottom of the brick. The loss is probably not worth the gain.