Improved Font Shrinking for mobile
Fonts will shrink if a word is too large to fit across a screen size until it matches the width perfectly. This improves upon the function that previously did this by wrapping each word in a <span> tag, checking it's width, and resizing based on that width. Should be faster and more robust.