Bricks now get the Stack ID from themselves
Formaerly, A Stack ID was passed to the mp_brick function. Now, each Brick will get its Stack ID from its mp_stack_order_STACKID meta key through the database directly. It then saves that value under "mp_stack_id". New bricks will automatically save this value. This is done because way back at the start of MP Stacks, bricks could be in multiple Stacks at once. Now, A Brick can only be in 1 Stack. Because Brick could be in multiple Stacks at once, there was no point in saving a single Stack ID to go with each Brick. Now there is. So we are doing that for each Brick when it loads.