Added filter flag called "mp_stacks_execute_mp_brick_in_mp_stack" which allows you to check whether a Stack's output should be executed. For example, you might want to require a password before a Brick's/Stack's Content is executed, This fixes a bug where styles and scripts were still enqueued even if the HTML for the Brick was not.
Added filter flag called "mp_stacks_execute_mp_brick_in_mp_stack" which allows you to check whether a Stack's output should be executed. For example, you might want to require a password before a Brick's/Stack's Content is executed, This fixes a bug where styles and scripts were still enqueued even if the HTML for the Brick was not.